move types validation from runtime to test
