+++ api/api.yaml
@@ -31,7 +31,7 @@ paths:
         name: id
         required: true
         schema:
-          enum: ["Header","Block","Extrinsic","EpochMarker","AvailabilityAssurance","AssurancesExtrinsic","Culprit","Fault","Judgement","Verdict","DisputesExtrinsic","Credential","ReportGuarantee","GuaranteesExtrinsic","Preimage","PreimageExtrinsic","RefineContext","SignedTicket","Ticket","TicketExtrinsic","ImportSpec","WorkItem","WorkItemExtrinsicSpec","WorkPackage","WorkPackageSpec","WorkReport","WorkExecResult","WorkResult"]
+          enum: ["Header","Block", "Extrinsic", "EpochMarker","AvailabilityAssurance","AssurancesExtrinsic","Culprit","Fault","Judgement","Verdict","DisputesExtrinsic","Credential","ReportGuarantee","GuaranteesExtrinsic","Preimage","PreimageExtrinsic","RefineContext","SignedTicket","Ticket","TicketExtrinsic","ImportSpec","WorkItem","WorkItemExtrinsicSpec","WorkPackage","WorkPackageSpec","WorkReport","WorkExecResult","WorkResult"]
           type: string   
       requestBody:
         content:
