+++ api/src/index.ts
@@ -5,10 +5,8 @@ import path from "node:path";
 import { fastifySwagger } from "@fastify/swagger";
 import { fastifySwaggerUi } from "@fastify/swagger-ui";
 
-import { areArraysEqual } from "./array-utils";
 import { codecToJson } from "./codec-to-json";
 import { jsonToCodec } from "./json-to-codec";
-import { kindNames } from "./kinds";
 import { getTypes } from "./types";
 
 const HOST = process.env.HOST || "127.0.0.1";
@@ -21,7 +19,7 @@ const start = async () => {
   await fastify.register(import("fastify-raw-body"), {
     field: "rawBody",
     global: true,
-    encoding: "utf8", // set it to false to set rawBody as a Buffer **Default utf8**
+    encoding: "utf8",
     runFirst: true,
   });
 
@@ -30,25 +28,6 @@ const start = async () => {
     specification: {
       path: "./api.yaml",
       baseDir: __dirname,
-      postProcessor: (spec) => {
-        for (const path of Object.values(spec.paths)) {
-          for (const properties of Object.values(path)) {
-            if (!properties.parameters) {
-              continue;
-            }
-            for (const param of properties.parameters) {
-              if (
-                param.description === "Type name" &&
-                param.in === "path" &&
-                !areArraysEqual(param.schema.enum, kindNames)
-              ) {
-                throw new Error("Type lists in kind.ts and api.yaml are different!");
-              }
-            }
-          }
-        }
-        return spec;
-      },
     },
   });
